Fixed License Headers and Finished last Catalog Types
