+++ src/main/java/org/spongepowered/common/mixin/core/stats/MixinAchievement.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.mixin.core.data.types;
+package org.spongepowered.common.mixin.core.stats;
 
 import net.minecraft.stats.IStatType;
 import net.minecraft.stats.StatBase;
